Moving template to own file. Some refactoring
