+++ src/checker.js
@@ -4,6 +4,8 @@ const fs = require('fs');
 const readline = require('readline');
 const ejs = require('ejs');
 const minimist = require('minimist');
+const template = require('./templates/scan-result');
+
 
 const appArgumentsDesc = `
   Usage: node checker.js --crawlFilePath <string> --filePrefix <string>
@@ -50,7 +52,7 @@ const runScan = async (page, urlFromFile, resultFolderPath, filePrefix, urlCount
   console.log('injected. get results');
 
   // Get the results from `axe.run()`.
-  results = await handle.jsonValue();
+  const results = await handle.jsonValue();
   console.log('write results to file');
 
   const resultFile = `${resultFolderPath}${filePrefix}_${urlCounter}.json`;
@@ -59,45 +61,7 @@ const runScan = async (page, urlFromFile, resultFolderPath, filePrefix, urlCount
   const data = JSON.stringify(results);
   fs.writeFileSync(resultFile, data);
 
-  let html = ejs.render(`<html><head>
-    <style>
-      .show {
-        display: block !important;
-      }
-    </style>
-    <script>
-      function toggle(id) {
-        document.getElementById(id).classList.toggle('show');
-      }
-    </script>
-  </head><body style="font-family:Arial;padding:20px;">
-  <h1>Results for <%=url %></h1>
-  <p>Total violations for this webpage: <%=results.violations.length %></p>
- 
-  <h2>Violation Details</h2>
-    <% results.violations.forEach(function(violation, index){ %>
-        <div style="border:1px solid grey;
-        padding:15px;
-        border-radius:5px;margin-bottom:15px;">
-          <h3><%=index+1 %>. <%=violation.description%></h3>
-          <h4>Severity: <span style="color:<%=violation.impact == 'minor' ? 'green': 'red'%>;" ><%=violation.impact %></span></h4>
-          <h5>Tags: 
-            <% violation.tags.forEach(function(tag){ %>
-              <span style="margin-right:10px"><%=tag%></span>
-            <% }); %> 
-          </h5>
-          <ol type="a">
-            <% violation.nodes.forEach(function(node, idx){ %>
-              <li><p><%=node.failureSummary %></p>
-                <button id="btn-<%=index%>-<%=idx %>" type="button" onClick="toggle('code-<%=index%>-<%=idx %>')">Show Target Markup</button>
-                <pre id="code-<%=index%>-<%=idx %>" style="display:none;white-space:break-spaces;overflow-x:auto;"><%=node.html %></pre>
-              </li>
-            <% }); %>
-          </ol>
-        </div>
-    <% }); %>
- 
-</body></html>`, { results: results, url: urlFromFile });
+  const html = ejs.render(template.resultTemplate, { results: results, url: urlFromFile });
 
   fs.writeFileSync(`${resultFolderPath}${filePrefix}_${urlCounter}.html`, html);
 
@@ -109,11 +73,11 @@ const injectAxe = (page) => {
     // Inject axe source code
     ${axeCore.source}
     // Run axe
-    // axe.configure({
-    //   rules: [{
-    //     id: 'tagRule',
-    //     tags: ['wcag2a']
-    //   }]
+    // axe.run({
+    //   runOnly: {
+    //     type: 'tag',
+    //     values: ['wcag2a']
+    //   }
     // })
     axe.run()
   `);
@@ -146,7 +110,7 @@ const runChecker = (crawlFilePath, filePrefix) => {
 
 exports.launch = (() => {
   let validArgs = true;
-  let argv = minimist(process.argv.slice(2));
+  const argv = minimist(process.argv.slice(2));
 
   const crawlFilePath = argv.crawlFilePath;
   const filePrefix = argv.filePrefix;
